Enhance deployment scripts and configuration for improved stability and security
- Update Dockerfile to include openssl-dev for PHP extensions
- Ensure JWT TTL and refresh TTL are integers in jwt.php
- Simplify deploy.sh by removing unnecessary functions and improving deployment steps
- Update docker-compose files to include JWT_SECRET and Redis configurations
- Modify www.conf to listen on 9000 directly
- Force migration execution in run_migrations.sh for reliability

+++ config/jwt.php
@@ -101,7 +101,8 @@ return [
     |
     */
 
-    'ttl' => env('JWT_TTL', 60),
+    // Ensure TTL is an integer to avoid type errors when used with Carbon
+    'ttl' => (int) env('JWT_TTL', 60),
 
     /*
     |--------------------------------------------------------------------------
@@ -120,7 +121,8 @@ return [
     |
     */
 
-    'refresh_ttl' => env('JWT_REFRESH_TTL', 20160),
+    // Ensure refresh TTL is an integer as well
+    'refresh_ttl' => (int) env('JWT_REFRESH_TTL', 20160),
 
     /*
     |--------------------------------------------------------------------------
